@@ -1,9 +1,11 @@
 name: flutter_jscore_example
 description: Demonstrates how to use the flutter_jscore plugin.
 publish_to: 'none'
+version: '1.0.0'
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
+  flutter: ">=1.24.0"
 
 dependencies:
   flutter:
