@@ -6,13 +6,14 @@ repository: https://github.com/xuelongqy/flutter_jscore
 issue_tracker: https://github.com/xuelongqy/flutter_jscore/issues
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
+  flutter: ">=1.24.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
-  ffi: ^0.1.3
+  ffi: ^1.0.0
 
 dev_dependencies:
   flutter_test:
