@@ -1,8 +1,8 @@
 import 'dart:ffi';
 import 'dart:io';
 
-final DynamicLibrary jscLib = Platform.isAndroid
-    ? DynamicLibrary.open("libjsc.so")
-    : Platform.isIOS || Platform.isMacOS
-        ? DynamicLibrary.open("JavaScriptCore.framework/JavaScriptCore")
-        : null;
+final DynamicLibrary jscLib = Platform.isIOS || Platform.isMacOS
+    ? DynamicLibrary.open("JavaScriptCore.framework/JavaScriptCore")
+    : Platform.isWindows
+        ? DynamicLibrary.open("libjsc.dll")
+        : DynamicLibrary.open("libjsc.so");
